style: Slick css global로 이동
