+++ client/src/styles/Global.jsx
@@ -50,6 +50,45 @@ const globalStyles = theme => css`
   li {
     list-style: none;
   }
+
+  // Slick
+
+  .slick-list {
+    overflow: hidden;
+  }
+  .slick-track {
+    display: flex;
+    align-items: center;
+  }
+  .slick-slide {
+  }
+
+  .slick-dots {
+    display: flex;
+    justify-content: center;
+
+    & > *:not(:last-child) {
+      margin-right: 0.7em;
+    }
+
+    .slick-active {
+      & span {
+        background: ${theme.colors.text.primary};
+        border-radius: 10px;
+      }
+    }
+  }
+
+  .dots__dot {
+    & > span {
+      display: inline-block;
+      width: 0.8em;
+      height: 0.8em;
+      border-radius: 50%;
+      background: ${theme.colors.text.disable_placeholder};
+      transition: all 0.3s ease-in-out;
+    }
+  }
 `;
 
 const GlobalStyles = ({ theme }) => <Global styles={globalStyles(theme)} />;
